Remove the "special" nature of publish-utils' name property.
